% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{adjustPvalues}
\alias{adjustPvalues}
\title{Adjust p values for multiple testing}
\usage{
adjustPvalues(result_loci_regr_df, method)
}
\arguments{
\item{result_loci_regr_df}{A data.frame of uncorrected p values for regressions for each CpG}

\item{method}{Correction method, a character string. Can be abbreviated.}
}
\value{
A data.frame of uncorrected and corrected p values for regressions for each CpG
}
\description{
Adjust p values for multiple testing
}
\seealso{
\code{\link[=p.adjust]{p.adjust()}} for p values adjustment method
}
