% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RLM-CaseControl.R
\name{casectrlAnalysis}
\alias{casectrlAnalysis}
\title{Run differential analysis using limma: CpG ~ Treatment + Confounders
Warning: does not allow missing data in confounders}
\usage{
casectrlAnalysis(
  meth_data,
  covariates,
  treatment,
  clinical_confounders = 0,
  technical_confounders = 0,
  transformToMvalue = FALSE,
  paired = FALSE,
  pair.lab = "pair",
  robust = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{meth_data}{A numeric matrix containing methylation data (CpGs as rows, samples as columns)}

\item{covariates}{A data.frame containing all variables needed for regression (exposures and confounders)}

\item{treatment}{A character string with the status name (must be a column in covariates)}

\item{clinical_confounders}{A character vector defining clinical confounders the regression will be adjusted for (must be columns in covariates)}

\item{technical_confounders}{A character vector defining technical confounders the regression will be adjusted for (must be columns in covariates)}

\item{transformToMvalue}{Boolean. Whether to transform CpG values into M-values. Default: FALSE.}

\item{paired}{Boolean. Whether samples are paired. Default: FALSE.}

\item{pair.lab}{String. Column name were to find pairing. Default: "pair".}

\item{robust}{Boolean. Whether to perform robust linear regression. Default: FALSE.}

\item{verbose}{Boolean. Whether to output messages. Default: TRUE.}

\item{maxit}{The number of iterations for each robust regression.}
}
\value{
An object of class MArrayLM as output by the limma::eBayes() function.
}
\description{
Run differential analysis using limma: CpG ~ Treatment + Confounders
Warning: does not allow missing data in confounders
}
