% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LocusWiseMultipleExposures.R
\name{LocusWiseMultiRLM}
\alias{LocusWiseMultiRLM}
\title{Robust linear regression for mean methylation level in function of multiple
exposure: methylation ~ exposure1 + exposure2 + exposure3 + W + epsilon}
\usage{
LocusWiseMultiRLM(
  meth_data,
  covariates,
  exposures,
  study_name,
  clinical_confounders,
  technical_confounders,
  transformToMvalue = FALSE,
  maxit,
  ncores = 1,
  path,
  file_name
)
}
\arguments{
\item{meth_data}{A matrix containing methylation data, CpG in rows.}

\item{covariates}{A data.frame containing all variables needed for regression (exposures and confounders)}

\item{exposures}{A character vector naming the exposures}

\item{study_name}{A name for the study.}

\item{clinical_confounders}{A character vector containing names of the confounders}

\item{technical_confounders}{A character vector defining technical confounders the regression will be adjusted to}

\item{transformToMvalue}{Boolean: whether input data should be transformed to Mvalue (ie: logit transformation)}

\item{maxit}{The number of iterations for each robust regression}

\item{ncores}{The number of cores used for parallel computing, by default all available cores}

\item{path}{Path for saving the result file}

\item{file_name}{File name without extension}
}
\value{
A data.frame with annotated statistics for each CpG: beta estimate for the association with exposure; uncorrected and corrected p values; average mean meth level with SE;
}
\description{
Robust linear regression for mean methylation level in function of multiple
exposure: methylation ~ exposure1 + exposure2 + exposure3 + W + epsilon
}
