% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LocusWiseRegression.R
\name{LocusWiseTimeMLReg}
\alias{LocusWiseTimeMLReg}
\title{Results of the linear regressions for mean methylation level in function of exposure}
\usage{
LocusWiseTimeMLReg(
  meth_data,
  covariates.df,
  exposures,
  time,
  clinical_confounders,
  technical_confounders,
  transformToMvalue = FALSE,
  ncores = bigstatsr::nb_cores(),
  path,
  file_name
)
}
\arguments{
\item{meth_data}{A matrix containing methylation data}

\item{exposures}{A character vector naming the exposures}

\item{technical_confounders}{A character vector defining technical confounders the regression will be adjusted to}

\item{transformToMvalue}{Boolean: whether input data should be transformed to Mvalue}

\item{ncores}{The number of cores used for parallel computing, by default all available cores}

\item{path}{Path for saving the result file}

\item{file_name}{File name without extension}

\item{expo_cov}{A data.frame containing all variables needed for regression (exposures and confounders)}

\item{confounders}{A character vector containing names of the confounders}

\item{annotation_object}{An Illumina object used to annotate files}

\item{expo_labels}{A character vector containing full names of exposures}
}
\value{
A data.frame with annotated statistics for each CpG: beta estimate for the association with exposure; uncorrected and corrected p values; average mean meth level with SE;
}
\description{
Results of the linear regressions for mean methylation level in function of exposure
}
