% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLM-2S.R
\name{LocusWiseTwoStageMLM}
\alias{LocusWiseTwoStageMLM}
\title{Results of the mixed linear regressions for mean methylation level in function of repeated exposure measurements}
\usage{
LocusWiseTwoStageMLM(
  meth_data,
  covariates,
  exposures,
  clinical_confounders,
  technical_confounders,
  transformToMvalue = FALSE,
  robust = TRUE,
  maxit,
  ncores = ceiling(bigstatsr::nb_cores()/2),
  path,
  file_name
)
}
\arguments{
\item{meth_data}{A matrix containing methylation data (samples in columns).}

\item{covariates}{A data.frame containing all variables needed for regression (exposures and confounders).}

\item{exposures}{A character vector naming the exposures.}

\item{clinical_confounders}{A character vector containing names of the confounders.}

\item{technical_confounders}{A character vector defining technical confounders the regression will be adjusted for.}

\item{transformToMvalue}{Boolean. Whether input data should be transformed to M-value}

\item{robust}{Boolean. Whether to perform robust linear regressions in the second stage. Default: TRUE.}

\item{maxit}{The number of iterations for each robust regression.}

\item{ncores}{The number of cores used for parallel computing, by default half of all available cores.}

\item{path}{Path for saving the result file.}

\item{file_name}{File name without extension.}
}
\value{
A data.frame with annotated statistics for each CpG: beta estimate for the association with exposure; uncorrected and corrected p values; average mean meth level with SE;
}
\description{
Results of the mixed linear regressions for mean methylation level in function of repeated exposure measurements
}
