% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{m2beta}
\alias{m2beta}
\title{Transform M-value to beta-value}
\usage{
m2beta(m)
}
\arguments{
\item{m}{A numeric value (eg: M-value)}
}
\value{
2^m/(2^m + 1)
}
\description{
Transform M-value to beta-value
}
