% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLM-2S.R
\name{MLM2_stage1}
\alias{MLM2_stage1}
\title{Run Mixed Linear regression for estimating average exposure level
(when repeatedly measured)}
\usage{
MLM2_stage1(exposure, time = NULL, data)
}
\arguments{
\item{exposure}{A character string with the exposure name}

\item{time}{A character string with the time variable name (for fitting autocorrelation AR1 model). Default: NULL}

\item{data}{A data.frame containing all variables needed for regression (exposures)}
}
\value{
A 2-column data.frame: id x average exposure.
}
\description{
Run Mixed Linear regression for estimating average exposure level
(when repeatedly measured)
}
