% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLM-2S.R
\name{MLM2_stage2}
\alias{MLM2_stage2}
\title{Run Mixed Linear regression for methylation at specific loci ~ 1 exposure at a time (ExWAS) adjusted}
\usage{
MLM2_stage2(
  CpG,
  covariates,
  exposure,
  clinical_confounders = 0,
  technical_confounders = 0,
  transformToMvalue = transformToMvalue,
  robust = TRUE,
  maxit = 25
)
}
\arguments{
\item{CpG}{A numeric vector containing methylation data from one CpG}

\item{covariates}{A data.frame containing all variables needed for regression (exposures and confounders)}

\item{exposure}{A character string with the exposure name}

\item{clinical_confounders}{A character vector containing names of the confounders}

\item{technical_confounders}{A character vector defining technical confounders the regression will be adjusted for}

\item{transformToMvalue}{Boolean: whether input data should be transformed to M-value}

\item{robust}{Boolean. Whether to perform robust linear regressions in the second stage. Default: TRUE.}

\item{maxit}{The number of iterations for each robust regression}
}
\value{
A named vector of p values for regressions for each CpG
}
\description{
Run Mixed Linear regression for methylation at specific loci ~ 1 exposure at a time (ExWAS) adjusted
}
