% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LocusWiseMultipleExposures.R
\name{multiRobustLinearRegression}
\alias{multiRobustLinearRegression}
\title{Run robust linear regression for methylation at specific loci ~ exposures}
\usage{
multiRobustLinearRegression(
  CpG,
  covariate_data,
  exposures,
  clinical_confounders = 0,
  technical_confounders = 0,
  transformToMvalue = FALSE,
  maxit
)
}
\arguments{
\item{CpG}{A numeric vector containing methylation data from 1 CpG}

\item{covariate_data}{A data.frame containing all variables needed for regression (exposures and confounders)}

\item{exposures}{A character string with the exposure name(s)}

\item{clinical_confounders}{A character vector containing names of the confounders}

\item{technical_confounders}{A character vector defining technical confounders the regression will be adjusted to}

\item{transformToMvalue}{Boolean. Whether CpG values should be turned to M-values.}

\item{maxit}{The number of iterations for each robust regression}
}
\value{
A named vector of p values for regressions for each CpG
}
\description{
Run robust linear regression for methylation at specific loci ~ exposures
}
