% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{myLinearRegressionLoci}
\alias{myLinearRegressionLoci}
\title{Run linear regression for methylation at specific loci ~ 1 exposure at a time (ExWAS) adjusted for diferent set of confounders}
\usage{
myLinearRegressionLoci(
  CpG,
  exposure,
  expo_cov,
  confounders = 0,
  technical_confounders = 0,
  transformToMvalue = transformToMvalue
)
}
\arguments{
\item{CpG}{A numeric vector containing methylation data from 1 CpG}

\item{exposure}{A character string with the exposure name}

\item{expo_cov}{A data.frame containing all variables needed for regression (exposures and confounders)}

\item{confounders}{A character vector containing names of the confounders}

\item{technical_confounders}{A character vector defining technical confounders the regression will be adjusted to}

\item{maxit}{The number of iterations for each robust regression}
}
\value{
A named vector of p values for regressions for each CpG
}
\description{
Run linear regression for methylation at specific loci ~ 1 exposure at a time (ExWAS) adjusted for diferent set of confounders
}
