% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLM-1S.R
\name{myMLRegressionLoci.Wald}
\alias{myMLRegressionLoci.Wald}
\title{Run Mixed Linear regression for methylation at a specific locus and repeated
exposure. Wald test.}
\usage{
myMLRegressionLoci.Wald(
  CpG,
  exposure,
  covariates,
  clinical_confounders = 0,
  technical_confounders = 0,
  transformToMvalue = transformToMvalue
)
}
\arguments{
\item{CpG}{A numeric vector containing methylation data from 1 CpG}

\item{exposure}{A character string with the exposure name}

\item{covariates}{A data.frame containing all variables needed for regression (exposures and confounders)}

\item{clinical_confounders}{A character vector defining clinical confounders the regression will be adjusted to}

\item{technical_confounders}{A character vector defining technical confounders the regression will be adjusted to}
}
\value{
A named vector of p values for regressions for each CpG
}
\description{
Run Mixed Linear regression for methylation at a specific locus and repeated
exposure. Wald test.
}
