% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{myObtainStatistics}
\alias{myObtainStatistics}
\title{Obtain regression statistics for one}
\usage{
myObtainStatistics(fit, data, exposure)
}
\arguments{
\item{fit}{An object created by MASS:mlr robust linear regression function}

\item{data}{A data.frame containing a subset of methylation data of one CpG (y) and exposure-covariates data (x)}

\item{exposure}{A string with the exposure name}
}
\value{
A data.frame containing results of the regression for one model for one exposure
}
\description{
Obtain regression statistics for one
}
