% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{myRobustLinearRegression}
\alias{myRobustLinearRegression}
\title{Fit the robust linear regression}
\usage{
myRobustLinearRegression(formula, data, maxit)
}
\arguments{
\item{formula}{A formula defining the elements of the regression}

\item{data}{A data.frame containing global methylation variable, exposure variables, confounders and technical confounders}

\item{maxit}{The number of iterations for each robust regression}
}
\value{
An object created by MASS:mlr robust linear regression function
}
\description{
Fit the robust linear regression
}
