% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LocusWiseCaseControl.R
\name{myTTest}
\alias{myTTest}
\title{Student T-Test}
\usage{
myTTest(formula, data, paired = FALSE)
}
\arguments{
\item{formula}{A formula defining the elements of the regression}

\item{data}{A data.frame containing global methylation variable, exposure variables, confounders and technical confounders}
}
\value{
An object created by lm linear regression function
}
\description{
Student T-Test
}
