% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RLM.R
\name{RLM}
\alias{RLM}
\title{Run robust linear regression: CpG ~ Exposure + Confounders}
\usage{
RLM(
  CpG,
  covariates,
  exposure,
  clinical_confounders = 0,
  technical_confounders = 0,
  maxit,
  transformToMvalue = FALSE
)
}
\arguments{
\item{CpG}{A numeric vector containing methylation data from 1 CpG}

\item{covariates}{A data.frame containing all variables needed for regression (exposures and confounders)}

\item{exposure}{A character string with the exposure name (must be a column in covariates)}

\item{clinical_confounders}{A character vector defining clinical confounders the regression will be adjusted for (must be columns in covariates)}

\item{technical_confounders}{A character vector defining technical confounders the regression will be adjusted for (must be columns in covariates)}

\item{maxit}{The number of iterations for each robust regression}

\item{transformToMvalue}{Boolean. Whether to transform CpG values into M-values. Default: FALSE.}
}
\value{
A named vector of p values for regressions for each CpG
}
\description{
Run robust linear regression: CpG ~ Exposure + Confounders
}
