% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Limma.R
\name{runLimma}
\alias{runLimma}
\title{Run linear regression using limma: CpG ~ Exposure + Confounders
Warning: does not allow missing data in confounders/exposures}
\usage{
runLimma(
  meth_data,
  covariates,
  exposure,
  clinical_confounders = 0,
  technical_confounders = 0,
  transformToMvalue = FALSE,
  robust = FALSE,
  verbose = TRUE,
  id.lab = "id"
)
}
\arguments{
\item{meth_data}{A numeric matrix containing methylation data (CpGs as rows, samples as columns)}

\item{covariates}{A data.frame containing all variables needed for regression (exposures and confounders)}

\item{exposure}{A character string with the exposure name (must be a column in covariates)}

\item{clinical_confounders}{A character vector defining clinical confounders the regression will be adjusted for (must be columns in covariates)}

\item{technical_confounders}{A character vector defining technical confounders the regression will be adjusted for (must be columns in covariates)}

\item{transformToMvalue}{Boolean. Whether to transform CpG values into M-values. Default: FALSE.}

\item{robust}{Boolean. Whether to perform robust linear regression. Default: FALSE.}

\item{verbose}{Boolean. Whether to output messages. Default: TRUE.}

\item{id.lab}{String. Column name were to find sample ids. Default: "id".}

\item{maxit}{The number of iterations for each robust regression.}
}
\value{
An object of class MArrayLM as output by the limma::eBayes() function.
}
\description{
Run linear regression using limma: CpG ~ Exposure + Confounders
Warning: does not allow missing data in confounders/exposures
}
