% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{synchronizeDatasets}
\alias{synchronizeDatasets}
\title{Synchronize methylation and meta datasets (eg: by trimester)}
\usage{
synchronizeDatasets(mmList, rm_samples = NULL)
}
\arguments{
\item{mmList}{A 2-entry list where mmList[\link{1}] is a matrix containing full methylation data
and mmList[\link{2}] is a data.frame containing all covariates}

\item{rm_samples}{IDs of samples to be removed from analyses (eg: spurious ones, outliers)}
}
\value{
A 2-entry list with synchronized data (ie: same samples, matching order)
}
\description{
Synchronize methylation and meta datasets (eg: by trimester)
}
