% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{TimeMLReg}
\alias{TimeMLReg}
\title{Run Mixed Linear regression for methylation at specific loci ~ 1 exposure at a time (ExWAS) adjusted}
\usage{
TimeMLReg(
  CpG,
  covariates.df,
  mainEffect,
  time,
  testTimeInteraction = TRUE,
  threshold.interaction = 0.05,
  clinical_confounders = 0,
  technical_confounders = 0,
  transformToMvalue = transformToMvalue,
  plot = FALSE,
  linearity.check = FALSE,
  plot.points = FALSE
)
}
\arguments{
\item{CpG}{A numeric vector containing methylation data from 1 CpG}

\item{technical_confounders}{A character vector defining technical confounders the regression will be adjusted to}

\item{expo_cov}{A data.frame containing all variables needed for regression (exposures and confounders)}

\item{exposure}{A character string with the exposure name}

\item{confounders}{A character vector containing names of the confounders}
}
\value{
A named vector of p values for regressions for each CpG
}
\description{
Run Mixed Linear regression for methylation at specific loci ~ 1 exposure at a time (ExWAS) adjusted
}
